Adding link to 2nd manual install script (geth-lodestar)
For some time, the only instructions to manually "run your own node" on Ephemery were for the geth-teku client combination ("geku"). These need to be either updated or retired, in favour of more recent contributions.

Further to the contribution of `setup-geth-lodestar.md` by @atkinsonholly this PR adds a link to these instructions from the "main ephemery onboarding flow", making them the primary route for running your own node on Ephemery.